requirements.txt
# Core
python-dotenv>=1.0.0

# Web API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
openai>=1.40.0

# RAG / ML
numpy>=1.24.0
faiss-cpu>=1.8.0
pymupdf>=1.24.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
